--- src/FairDispatchable.php
@@ -1,132 +1,0 @@
-<?php
-
-namespace Aloware\TenantsQueue;
-
-use Illuminate\Contracts\Bus\Dispatcher;
-use Illuminate\Support\Fluent;
-use Illuminate\Support\Facades\Redis;
-trait FairDispatchable
-{
-    public $tenantId;
-
-    /**
-     * Dispatch the job with the given arguments.
-     *
-     * @param  mixed  ...$arguments
-     * @return \Illuminate\Foundation\Bus\PendingDispatch
-     */
-    public static function dispatch(...$arguments)
-    {
-        return new PendingDispatch(new static(...$arguments));
-    }
-
-    /**
-     * Dispatch the job with the given arguments if the given truth test passes.
-     *
-     * @param  bool|\Closure  $boolean
-     * @param  mixed  ...$arguments
-     * @return \Illuminate\Foundation\Bus\PendingDispatch|\Illuminate\Support\Fluent
-     */
-    public static function dispatchIf($boolean, ...$arguments)
-    {
-        if ($boolean instanceof Closure) {
-            $dispatchable = new static(...$arguments);
-
-            return value($boolean, $dispatchable)
-                ? new PendingDispatch($dispatchable)
-                : new Fluent;
-        }
-
-        return value($boolean)
-            ? new PendingDispatch(new static(...$arguments))
-            : new Fluent;
-    }
-
-    /**
-     * Dispatch the job with the given arguments unless the given truth test passes.
-     *
-     * @param  bool|\Closure  $boolean
-     * @param  mixed  ...$arguments
-     * @return \Illuminate\Foundation\Bus\PendingDispatch|\Illuminate\Support\Fluent
-     */
-    public static function dispatchUnless($boolean, ...$arguments)
-    {
-        if ($boolean instanceof Closure) {
-            $dispatchable = new static(...$arguments);
-
-            return ! value($boolean, $dispatchable)
-                ? new PendingDispatch($dispatchable)
-                : new Fluent;
-        }
-
-        return ! value($boolean)
-            ? new PendingDispatch(new static(...$arguments))
-            : new Fluent;
-    }
-
-    /**
-     * Dispatch a command to its appropriate handler in the current process.
-     *
-     * Queueable jobs will be dispatched to the "sync" queue.
-     *
-     * @param  mixed  ...$arguments
-     * @return mixed
-     */
-    public static function dispatchSync(...$arguments)
-    {
-        return app(Dispatcher::class)->dispatchSync(new static(...$arguments));
-    }
-
-    /**
-     * Dispatch a command to its appropriate handler in the current process.
-     *
-     * @return mixed
-     *
-     * @deprecated Will be removed in a future Laravel version.
-     */
-    public static function dispatchNow(...$arguments)
-    {
-        return app(Dispatcher::class)->dispatchNow(new static(...$arguments));
-    }
-
-    /**
-     * Dispatch a command to its appropriate handler after the current process.
-     *
-     * @param  mixed  ...$arguments
-     * @return mixed
-     */
-    public static function dispatchAfterResponse(...$arguments)
-    {
-        return app(Dispatcher::class)->dispatchAfterResponse(new static(...$arguments));
-    }
-
-    /**
-     * Set the jobs that should run if this job is successful.
-     *
-     * @param  array  $chain
-     * @return \Illuminate\Foundation\Bus\PendingChain
-     */
-    public static function withChain($chain)
-    {
-        return new PendingChain(static::class, $chain);
-    }
-
-    /**
-     * Set the tenant stamping logic on the jobs.
-     *
-     * @return $this
-     */
-    public function tenant()
-    {
-        $redisDriver = config('tenants-queue.redis_driver');
-        $redis = Redis::connection($redisDriver);
-        $tenantsQueuePrefix = config('tenants-queue.prefix');
-        $tenantsSize = config('tenants-queue.tenants_size');
-        $tenantId = rand(0, $tenantsSize);
-        $listKeyName = "{$tenantsQueuePrefix}:{$this->queue}";
-        $redis->sadd($listKeyName, $tenantId);
-        $this->tenantId = $tenantId;
-        return $this;
-    }
-
-}
