+++ src/TenantsDispatchable.php
@@ -6,39 +6,107 @@ use Illuminate\Contracts\Bus\Dispatcher;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Support\Fluent;
 
-trait TenantsDispatchable
+trait TenantDispatchable
 {
-    use Dispatchable;
-
+    /**
+     * Dispatch the job with the given arguments.
+     *
+     * @param  mixed  ...$arguments
+     * @return \Illuminate\Foundation\Bus\PendingDispatch
+     */
     public static function dispatch(...$arguments)
     {
-
-        return new PendingDispatch($job);
+        return new TenantPendingDispatch(new static(...$arguments));
     }
 
+    /**
+     * Dispatch the job with the given arguments if the given truth test passes.
+     *
+     * @param  bool|\Closure  $boolean
+     * @param  mixed  ...$arguments
+     * @return \Illuminate\Foundation\Bus\PendingDispatch|\Illuminate\Support\Fluent
+     */
     public static function dispatchIf($boolean, ...$arguments)
     {
-        return $boolean ? new PendingDispatch($job) : new Fluent;
+        if ($boolean instanceof Closure) {
+            $dispatchable = new static(...$arguments);
+
+            return value($boolean, $dispatchable)
+                ? new TenantPendingDispatch($dispatchable)
+                : new Fluent;
+        }
+
+        return value($boolean)
+            ? new TenantPendingDispatch(new static(...$arguments))
+            : new Fluent;
     }
 
+    /**
+     * Dispatch the job with the given arguments unless the given truth test passes.
+     *
+     * @param  bool|\Closure  $boolean
+     * @param  mixed  ...$arguments
+     * @return \Illuminate\Foundation\Bus\PendingDispatch|\Illuminate\Support\Fluent
+     */
     public static function dispatchUnless($boolean, ...$arguments)
     {
-        return !$boolean ? new PendingDispatch($job) : new Fluent;
+        if ($boolean instanceof Closure) {
+            $dispatchable = new static(...$arguments);
+
+            return ! value($boolean, $dispatchable)
+                ? new TenantPendingDispatch($dispatchable)
+                : new Fluent;
+        }
+
+        return ! value($boolean)
+            ? new TenantPendingDispatch(new static(...$arguments))
+            : new Fluent;
     }
 
+    /**
+     * Dispatch a command to its appropriate handler in the current process.
+     *
+     * Queueable jobs will be dispatched to the "sync" queue.
+     *
+     * @param  mixed  ...$arguments
+     * @return mixed
+     */
     public static function dispatchSync(...$arguments)
     {
-        return app(Dispatcher::class)->dispatchSync($job);
+        return app(Dispatcher::class)->dispatchSync(new static(...$arguments));
     }
 
+    /**
+     * Dispatch a command to its appropriate handler in the current process.
+     *
+     * @return mixed
+     *
+     * @deprecated Will be removed in a future Laravel version.
+     */
     public static function dispatchNow(...$arguments)
     {
-        return app(Dispatcher::class)->dispatchNow($job);
+        return app(Dispatcher::class)->dispatchNow(new static(...$arguments));
     }
 
+    /**
+     * Dispatch a command to its appropriate handler after the current process.
+     *
+     * @param  mixed  ...$arguments
+     * @return mixed
+     */
     public static function dispatchAfterResponse(...$arguments)
     {
-        return app(Dispatcher::class)->dispatchAfterResponse($job);
+        return self::dispatch(...$arguments)->afterResponse();
     }
 
+    /**
+     * Set the jobs that should run if this job is successful.
+     *
+     * @param  array  $chain
+     * @return \Illuminate\Foundation\Bus\PendingChain
+     */
+    public static function withChain($chain)
+    {
+        return new PendingChain(static::class, $chain);
+    }
 }
